requirements.txt
python-telegram-bot==20.7
cohere==4.37
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0 